Crate checked

A safe API wrapper for Store and other utility crates that can be enabled through the features:

linker

interop

All extension methods defined in this module use special stored objects. These objects are essentially normal objects like FuncAddr, RunState or Value. However, they also contain an additional field of type StoreId as a tag to know to which Store they belong to.

While this is easy for address types like FuncAddr or MemAddr, some types are enums and their variants are visible to the user. For example, consider the Value enum, where users have full access to all of its variants. To be able to attach a tag only to the Value::Ref variant of this enum, the entire enum has to be re-defined. The result is a completely new type StoredValue.

Modules

interop 🔒

linker 🔒

store 🔒

stored_types 🔒

value 🔒

Structs

Linker

Store

StoreId

A unique identifier for a specific Store

Stored

A generic stored wrapper. This is used to wrap struct types such as FuncAddr, MemAddr, etc.

StoredHostCall

A stored variant of HostCall

StoredInstantiationOutcome

A stored variant of InstantiationOutcome

StoredRefFunc

Enums

StoredExternVal

A stored variant of ExternVal

StoredRef

A stored variant of Ref

StoredResumable

A stored variant of Resumable

StoredRunState

A stored variant of RunState

StoredValue

A stored variant of Value

Traits

AbstractStored

A trait for types that might have a StoreId attached to them, so-called stored types.

StoredInteropValue

A stored variant of InteropValue

StoredInteropValueList

A StoredInteropValueList is an iterable list of StoredInteropValues (i.e. Rust types that can be converted into Wasm StoredValues).